Move classes another package
