+++ client/src/main/java/connection/MessageOut.java
@@ -14,8 +14,7 @@ public class MessageOut implements Sender {
   }
 
   public static MessageOut from(Connector connector) throws IOException {
-    Boolean autoflush = true;
-    return new MessageOut(new PrintWriter(new OutputStreamWriter(connector.getOutStream(), StandardCharsets.UTF_8), autoflush));
+    return new MessageOut(new PrintWriter(new OutputStreamWriter(connector.getOutStream(), StandardCharsets.UTF_8)));
   }
 
   @Override
