+++ client/src/main/java/model/BoundariesPosition.java
@@ -1,5 +1,7 @@
 package model;
 
+import gui.fields.Mast;
+
 import java.util.ArrayList;
 import java.util.List;
